+++ src/oauth/index.ts
@@ -3,6 +3,9 @@ export * from "./oauth-token.interfaces";
 export * from "./oauth-token.manager";
 export * from "./oauth-crypto.utils";
 
+// Provider registry
+export * from "./oauth-provider.registry";
+
 // Token stores
 export * from "./stores/file-token.store";
 export * from "./stores/mongo-token.store";